@@ -8,8 +8,6 @@ from inspect import Signature
 import os
 import platform
 import re
-from typing import Optional
-from typing import Union
 
 import numpy as np
 from numpy.random import default_rng
@@ -279,17 +277,6 @@ def pytest_addoption(parser):
     parser.addoption('--test_downloads', action='store_true', default=False)
 
 
-def pytest_configure(config: pytest.Config):
-    """Add filterwarnings for vtk < 9.1 and numpy bool deprecation"""
-    warnings = config.getini('filterwarnings')
-
-    if pyvista.vtk_version_info < (9, 1):
-        warnings.append(
-            r'ignore:.*np\.bool.{1} is a deprecated alias for the builtin '
-            r'.{1}bool.*:DeprecationWarning'
-        )
-
-
 def _check_args_kwargs_marker(item_mark: pytest.Mark, sig: Signature):
     """Test for a given args and kwargs for a mark using its signature"""
 
@@ -355,39 +342,48 @@ def pytest_runtest_setup(item: pytest.Item):
 
     See custom marks in pyproject.toml.
     """
-
+    needs_vtk_version = 'needs_vtk_version'
     # this test needs a given VTK version
-    for item_mark in item.iter_markers('needs_vtk_version'):
+    for item_mark in item.iter_markers(needs_vtk_version):
         sig = Signature(
             [
                 Parameter(
                     'args',
                     kind=Parameter.VAR_POSITIONAL,
-                    annotation=Union[int, tuple[int]],
+                    annotation=int | tuple[int],
                 ),
                 Parameter(
                     'at_least',
                     kind=Parameter.KEYWORD_ONLY,
-                    annotation=Optional[tuple[int]],
+                    annotation=tuple[int] | None,
                     default=None,
                 ),
                 Parameter(
                     'less_than',
                     kind=Parameter.KEYWORD_ONLY,
                     default=None,
-                    annotation=Optional[tuple[int]],
+                    annotation=tuple[int] | None,
                 ),
                 Parameter(
                     'reason',
                     kind=Parameter.KEYWORD_ONLY,
                     default=None,
-                    annotation=Optional[str],
+                    annotation=str | None,
                 ),
             ]
         )
         _min, _max, bounds = _get_min_max_vtk_version(item_mark=item_mark, sig=sig)
         _min = (_min,) if isinstance(_min, int) else _min
         _max = (_max,) if isinstance(_max, int) else _max
+
+        if (_min is not None and _min <= pyvista._MIN_SUPPORTED_VTK_VERSION) or (
+            _max is not None and _max <= pyvista._MIN_SUPPORTED_VTK_VERSION
+        ):
+            msg = (
+                f'The {needs_vtk_version!r} marker is no longer necessary\n'
+                f'and can be removed from test {item}.'
+            )
+            raise pyvista.VTKVersionError(msg)
 
         curr_version = pyvista.vtk_version_info
 
@@ -460,13 +456,13 @@ def pytest_runtest_setup(item: pytest.Item):
                     p := 'processor',
                     kind=Parameter.KEYWORD_ONLY,
                     default=None,
-                    annotation=Union[str, None],
+                    annotation=str | None,
                 ),
                 Parameter(
                     m := 'machine',
                     kind=Parameter.KEYWORD_ONLY,
                     default=None,
-                    annotation=Union[str, None],
+                    annotation=str | None,
                 ),
             ]
         )
