@@ -14,10 +14,7 @@ from pyvista import examples
 from pyvista.plotting import charts
 from pyvista.plotting.colors import COLOR_SCHEMES
 
-pytestmark = [
-    pytest.mark.needs_vtk_version(9, 2),  # skip all tests if VTK<9.2.0
-    pytest.mark.skip_check_gc,  # A large number of tests here fail gc
-]
+pytestmark = pytest.mark.skip_check_gc  # A large number of tests here fail gc
 
 
 def vtk_array_to_tuple(arr):
